+++ src/main/java/com/carservice/dto/ServiceJobDTO.java
@@ -4,7 +4,9 @@ import com.carservice.data.entities.Appointment;
 import com.carservice.data.entities.Employee;
 import com.carservice.data.enums.JobState;
 import com.carservice.data.enums.JobType;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.AllArgsConstructor;
@@ -18,26 +20,26 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @ToString
 public class ServiceJobDTO {
-    private Long id;
+	private Integer id;
 
-    private Employee employee;
+	private Employee employee;
 
-    @Enumerated(EnumType.STRING)
-    @NotEmpty(message = "The type must be set!")
-    private JobType type;
+	@Enumerated(EnumType.STRING)
+	@NotEmpty(message = "The type must be set!")
+	private JobType type;
 
-    private LocalDate dateStarted;
+	private LocalDate dateStarted;
 
-    @Column(name = "date_finished")
-    private LocalDate dateFinished;
+	@Column(name = "date_finished")
+	private LocalDate dateFinished;
 
-    @PositiveOrZero(message = "The price can't be negative")
-    private BigDecimal price;
+	@PositiveOrZero(message = "The price can't be negative")
+	private BigDecimal price;
 
-    @Enumerated(EnumType.STRING)
-    @NotEmpty(message = "The state must be set!")
-    private JobState state = JobState.PENDING;
+	@Enumerated(EnumType.STRING)
+	@NotEmpty(message = "The state must be set!")
+	private JobState state = JobState.PENDING;
 
-    private Appointment appointment;
+	private Appointment appointment;
 
 }
