+++ src/main/java/com/carservice/repository/CarCenterRepository.java
@@ -6,9 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CarCenterRepository extends JpaRepository<CarCenter, Long> {
-    List<CarCenter> findAllByName(String name);
-    List<CarCenter> findByCarBrand(CarBrand carBrand);
+public interface CarCenterRepository extends JpaRepository<CarCenter, Integer> {
+	List<CarCenter> findAllByName(String name);
+
+	List<CarCenter> findByCarBrand(CarBrand carBrand);
 
 
 }
